Refactor: add texture.rs
