+++ lib/Cargo.toml
@@ -15,6 +15,7 @@ glam = "0.25.0"
 log = "0.4.22"
 wgpu = "22.0"
 winit = { version = "0.29.10", features = ["rwh_05"] }
+image = "0.25.2"
 
 [target.'cfg(windows)'.dependencies]
 winapi = { version = "0.3", features = ["winuser"] }
